+++ script/tool/lib/src/build_examples_command.dart
@@ -11,11 +11,21 @@ import 'common/core.dart';
 import 'common/package_looping_command.dart';
 import 'common/plugin_utils.dart';
 import 'common/process_runner.dart';
+import 'common/repository_package.dart';
 
 /// Key for APK.
 const String _platformFlagApk = 'apk';
 
-const int _exitNoPlatformFlags = 2;
+const int _exitNoPlatformFlags = 3;
+
+// Flutter build types. These are the values passed to `flutter build <foo>`.
+const String _flutterBuildTypeAndroid = 'apk';
+const String _flutterBuildTypeIos = 'ios';
+const String _flutterBuildTypeLinux = 'linux';
+const String _flutterBuildTypeMacOS = 'macos';
+const String _flutterBuildTypeWeb = 'web';
+const String _flutterBuildTypeWin32 = 'windows';
+const String _flutterBuildTypeWinUwp = 'winuwp';
 
 /// A command to build the example applications for packages.
 class BuildExamplesCommand extends PackageLoopingCommand {
@@ -29,6 +39,7 @@ class BuildExamplesCommand extends PackageLoopingCommand {
     argParser.addFlag(kPlatformMacos);
     argParser.addFlag(kPlatformWeb);
     argParser.addFlag(kPlatformWindows);
+    argParser.addFlag(kPlatformWinUwp);
     argParser.addFlag(kPlatformIos);
     argParser.addFlag(_platformFlagApk);
     argParser.addOption(
@@ -45,33 +56,40 @@ class BuildExamplesCommand extends PackageLoopingCommand {
     _platformFlagApk: const _PlatformDetails(
       'Android',
       pluginPlatform: kPlatformAndroid,
-      flutterBuildType: 'apk',
+      flutterBuildType: _flutterBuildTypeAndroid,
     ),
     kPlatformIos: const _PlatformDetails(
       'iOS',
       pluginPlatform: kPlatformIos,
-      flutterBuildType: 'ios',
+      flutterBuildType: _flutterBuildTypeIos,
       extraBuildFlags: <String>['--no-codesign'],
     ),
     kPlatformLinux: const _PlatformDetails(
       'Linux',
       pluginPlatform: kPlatformLinux,
-      flutterBuildType: 'linux',
+      flutterBuildType: _flutterBuildTypeLinux,
     ),
     kPlatformMacos: const _PlatformDetails(
       'macOS',
       pluginPlatform: kPlatformMacos,
-      flutterBuildType: 'macos',
+      flutterBuildType: _flutterBuildTypeMacOS,
     ),
     kPlatformWeb: const _PlatformDetails(
       'web',
       pluginPlatform: kPlatformWeb,
-      flutterBuildType: 'web',
+      flutterBuildType: _flutterBuildTypeWeb,
     ),
     kPlatformWindows: const _PlatformDetails(
-      'Windows',
+      'Win32',
       pluginPlatform: kPlatformWindows,
-      flutterBuildType: 'windows',
+      pluginPlatformVariant: platformVariantWin32,
+      flutterBuildType: _flutterBuildTypeWin32,
+    ),
+    kPlatformWinUwp: const _PlatformDetails(
+      'UWP',
+      pluginPlatform: kPlatformWindows,
+      pluginPlatformVariant: platformVariantWinUwp,
+      flutterBuildType: _flutterBuildTypeWinUwp,
     ),
   };
 
@@ -96,7 +114,7 @@ class BuildExamplesCommand extends PackageLoopingCommand {
   }
 
   @override
-  Future<PackageResult> runForPackage(Directory package) async {
+  Future<PackageResult> runForPackage(RepositoryPackage package) async {
     final List<String> errors = <String>[];
 
     final bool isPlugin = isFlutterPlugin(package);
@@ -109,8 +127,9 @@ class BuildExamplesCommand extends PackageLoopingCommand {
     // no package-level platform information for non-plugin packages.
     final Set<_PlatformDetails> buildPlatforms = isPlugin
         ? requestedPlatforms
-            .where((_PlatformDetails platform) =>
-                pluginSupportsPlatform(platform.pluginPlatform, package))
+            .where((_PlatformDetails platform) => pluginSupportsPlatform(
+                platform.pluginPlatform, package,
+                variant: platform.pluginPlatformVariant))
             .toSet()
         : requestedPlatforms.toSet();
     if (buildPlatforms.isEmpty) {
@@ -135,16 +154,16 @@ class BuildExamplesCommand extends PackageLoopingCommand {
     print('');
 
     bool builtSomething = false;
-    for (final Directory example in getExamplesForPackage(package)) {
+    for (final RepositoryPackage example in package.getExamples()) {
       final String packageName =
-          getRelativePosixPath(example, from: packagesDir);
+          getRelativePosixPath(example.directory, from: packagesDir);
 
       for (final _PlatformDetails platform in buildPlatforms) {
         // Repo policy is that a plugin must have examples configured for all
         // supported platforms. For packages, just log and skip any requested
         // platform that a package doesn't have set up.
         if (!isPlugin &&
-            !example
+            !example.directory
                 .childDirectory(platform.flutterPlatformDirectory)
                 .existsSync()) {
           print('Skipping ${platform.label} for $packageName; not supported.');
@@ -180,11 +199,27 @@ class BuildExamplesCommand extends PackageLoopingCommand {
   }
 
   Future<bool> _buildExample(
-    Directory example,
+    RepositoryPackage example,
     String flutterBuildType, {
     List<String> extraBuildFlags = const <String>[],
   }) async {
     final String enableExperiment = getStringArg(kEnableExperiment);
+
+    // The UWP template is not yet stable, so the UWP directory
+    // needs to be created on the fly with 'flutter create .'
+    Directory? temporaryPlatformDirectory;
+    if (flutterBuildType == _flutterBuildTypeWinUwp) {
+      final Directory uwpDirectory = example.directory.childDirectory('winuwp');
+      if (!uwpDirectory.existsSync()) {
+        print('Creating temporary winuwp folder');
+        final int exitCode = await processRunner.runAndStream(flutterCommand,
+            <String>['create', '--platforms=$kPlatformWinUwp', '.'],
+            workingDir: example.directory);
+        if (exitCode == 0) {
+          temporaryPlatformDirectory = uwpDirectory;
+        }
+      }
+    }
 
     final int exitCode = await processRunner.runAndStream(
       flutterCommand,
@@ -195,8 +230,15 @@ class BuildExamplesCommand extends PackageLoopingCommand {
         if (enableExperiment.isNotEmpty)
           '--enable-experiment=$enableExperiment',
       ],
-      workingDir: example,
+      workingDir: example.directory,
     );
+
+    if (temporaryPlatformDirectory != null &&
+        temporaryPlatformDirectory.existsSync()) {
+      print('Cleaning up ${temporaryPlatformDirectory.path}');
+      temporaryPlatformDirectory.deleteSync(recursive: true);
+    }
+
     return exitCode == 0;
   }
 }
@@ -206,6 +248,7 @@ class _PlatformDetails {
   const _PlatformDetails(
     this.label, {
     required this.pluginPlatform,
+    this.pluginPlatformVariant,
     required this.flutterBuildType,
     this.extraBuildFlags = const <String>[],
   });
@@ -215,6 +258,10 @@ class _PlatformDetails {
 
   /// The key in a pubspec's platform: entry.
   final String pluginPlatform;
+
+  /// The supportedVariants key under a plugin's [pluginPlatform] entry, if
+  /// applicable.
+  final String? pluginPlatformVariant;
 
   /// The `flutter build` build type.
   final String flutterBuildType;